application/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-cors-headers==3.1.1
django-cors-middleware
django-rest-swagger==2.2.0
django-summernote==0.8.11.4
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
pymysql==0.9.3
pillow==7.0.0
PyMySQL[rsa]
cryptography

django-allauth==0.40.0
django-bootstrap-modal-forms==1.5.0
django-bootstrap-pagination==1.7.1
django-bootstrap3==12.0.3
django-crispy-forms==1.8.1
django-datatables-view==1.19.1
django-filter==2.2.0
django-model-utils==4.0.0
django-social-auth==0.7.28
django-tables2==2.2.1
django-templated-mail==1.1.1
pandas
pyyaml

